moving cat list display
